+++ kaminarimon/backend.py
@@ -6,40 +6,64 @@ from django.conf import settings
 from django.contrib.auth import get_user_model
 from django.contrib.auth.backends import ModelBackend
 from django.contrib.auth.models import Group
+from ldap.filter import escape_filter_chars
 from rest_framework.exceptions import AuthenticationFailed
 
 
-# Global TODO: Move hardcoded data into settings
-def get_ldap_groups(user_dn):
-    # TODO: move into class and do connection only once on init
-    conn = ldap.initialize(settings.AUTH_LDAP_SERVER_URI, bytes_mode=False)
+LDAP_DEFAULT_SERVERS = [
+    "ldaps://s1.idm-001.prod.iad2.dc.redhat.com:636",
+    "ldaps://s2.idm-001.prod.iad2.dc.redhat.com:636",
+    "ldaps://s1.idm-001.prod.rdu2.dc.redhat.com:636",
+    "ldaps://s2.idm-001.prod.rdu2.dc.redhat.com:636",
+    "ldaps://s1.idm-001.prod.us-east-1.aws.redhat.com:636",
+    "ldaps://s2.idm-001.prod.us-east-1.aws.redhat.com:636",
+]
+LDAP_DEFAULT_BASE_DN = "dc=ipa,dc=redhat,dc=com"
+
+def _ldap_connect():
+    servers = getattr(settings, "KAMINARIMON_LDAP_SERVERS", LDAP_DEFAULT_SERVERS)
+
+    conn = ldap.initialize(" ".join(servers), bytes_mode=False)
     conn.set_option(ldap.OPT_X_TLS_REQUIRE_CERT, ldap.OPT_X_TLS_DEMAND)
-    conn.simple_bind_s("", "")
-    group_base = "ou=adhoc,ou=managedgroups,dc=redhat,dc=com"
-    filter = f"(&(objectClass=groupOfUniqueNames)(uniqueMember={user_dn}))"
-    attrlist = ["cn"]
-    groups = conn.search_s(group_base, ldap.SCOPE_SUBTREE, filter, attrlist)
+
+    bind_dn = getattr(settings, "KAMINARIMON_LDAP_BIND_DN", None)
+    if bind_dn:
+        # Simple bind for test/dev environments without Kerberos
+        bind_password = getattr(settings, "KAMINARIMON_LDAP_BIND_PASSWORD", "")
+        conn.simple_bind_s(bind_dn, bind_password)
+    else:
+        conn.sasl_gssapi_bind_s()
+
+    return conn
+
+# Global TODO: Move hardcoded data into settings
+def get_ldap_groups(user_dn, conn):
+    base_dn = getattr(settings, "LDAP_BASE_DN", LDAP_DEFAULT_BASE_DN)
+    group_base = f"cn=groups,cn=accounts,{base_dn}"
+    groups = conn.search_s(
+        group_base,
+        ldap.SCOPE_SUBTREE,
+        f"(member={escape_filter_chars(user_dn)})",
+        ["cn"],
+    )
     return {group[1]["cn"][0].decode() for group in groups}
 
 
-def get_user_info(username):
-    conn = ldap.initialize(settings.AUTH_LDAP_SERVER_URI, bytes_mode=False)
-    conn.set_option(ldap.OPT_X_TLS_REQUIRE_CERT, ldap.OPT_X_TLS_DEMAND)
-    conn.simple_bind_s("", "")
-    base = "ou=%s,dc=redhat,dc=com"
+def get_user_info(username, conn):
+    base_dn = getattr(settings, "LDAP_BASE_DN", LDAP_DEFAULT_BASE_DN)
+    user_base = f"cn=users,cn=accounts,{base_dn}"
     attrlist = ["givenName", "sn", "mail"]
     user = conn.search_s(
-        base % "users", ldap.SCOPE_SUBTREE, f"(uid={username})", attrlist
+        user_base,
+        ldap.SCOPE_SUBTREE,
+        f"(uid={escape_filter_chars(username)})",
+        attrlist
     )
     if not user:
-        user = conn.search_s(
-            base % "serviceaccounts", ldap.SCOPE_SUBTREE, f"(uid={username})", attrlist
+        # TODO: log mismatched username
+        raise AuthenticationFailed(
+            "Could not find matching LDAP account for Kerberos principal"
         )
-        if not user:
-            # TODO: log mismatched username
-            raise AuthenticationFailed(
-                "Could not find matching LDAP account for Kerberos principal"
-            )
     return user[0]
 
 
@@ -86,8 +110,14 @@ class LDAPRemoteUser(ModelBackend):
         Synchronize the user with the external system and return the updated user.
         """
         username = user.get_username()
-        dn, attrs = get_user_info(username)
-        groups = get_ldap_groups(dn)
+
+        conn = _ldap_connect()
+        try:
+            dn, attrs = get_user_info(username, conn)
+            groups = get_ldap_groups(dn, conn)
+        finally:
+            conn.unbind_s()
+
         # Note: we simply create Groups without handling Django-style permissions
         # since we don't really use them -- we use PostgreSQL RLS
         group_objs = [Group.objects.get_or_create(name=group)[0] for group in groups]